refactor: 헤더를 4개 항목으로 개편하고 예선 소개 영상 자동재생 적용
예선 진출팀·다시보기 섹션을 복원하고, 광탈페 예선(소개 영상·실시간
중계·공연 순서)과 광탈페 본선(위치·오시는 길)을 별도 섹션으로 분리해
헤더를 슬로건 포함 5개 항목으로 구성. 예선 소개 영상은 컨트롤 없이
자동 반복 재생되도록 변경.
