+++ src/widgets/main/PreliminaryLiveSection/index.tsx
@@ -1,0 +1,93 @@
+"use client";
+
+import Image from "next/image";
+import { useState } from "react";
+import { cn } from "@/shared/utils/cn";
+import { SectionTitle } from "@/shared/ui/SectionTitle";
+import YouTubeLazyEmbed from "@/shared/ui/YouTubeLazyEmbed";
+
+const LIVE_STREAMS = [
+  { date: "7월 24일 (금)", label: "1일차", videoId: "nL0pdTRB6Hc" },
+  { date: "7월 25일 (토)", label: "2일차", videoId: "eR3Tx134-dU" },
+] as const;
+
+const PreliminaryLiveSection = () => {
+  const [activeDay, setActiveDay] = useState(0);
+  const activeStream = LIVE_STREAMS[activeDay];
+
+  return (
+    <section className={cn("relative w-full overflow-hidden bg-white py-[80px] mobile:py-[52px]")}>
+      {/* 블러 원 데코 */}
+      <div
+        className="absolute top-[-60px] left-[-80px] w-[320px] h-[320px] rounded-full bg-orange-300 opacity-60 blur-[80px] pointer-events-none"
+        aria-hidden="true"
+      />
+      <div
+        className="absolute bottom-[-60px] right-[-80px] w-[280px] h-[280px] rounded-full bg-orange-400 opacity-50 blur-[80px] pointer-events-none"
+        aria-hidden="true"
+      />
+
+      <div className="relative z-10 mx-auto flex w-[70%] mobile:w-full flex-col gap-40 mobile:gap-28 px-16">
+        <SectionTitle
+          title="2026 광탈페 光트로 예선"
+          description={<>7월 24일(금)·25일(토), 광주학생교육문화회관에서 열립니다</>}
+        />
+
+        <div className="flex flex-col gap-12">
+          <h3 className="text-body2b mobile:text-body3b text-black">예선 소개 영상</h3>
+          <video
+            src="/video/preliminary-intro.mp4"
+            autoPlay
+            muted
+            loop
+            playsInline
+            preload="metadata"
+            className="w-full rounded-lg border border-gray-100 bg-black"
+          >
+            <track kind="captions" />
+          </video>
+        </div>
+
+        <div className="flex flex-col gap-16">
+          <h3 className="text-body2b mobile:text-body3b text-black">실시간 중계</h3>
+          <div className="flex gap-8">
+            {LIVE_STREAMS.map(({ date, label }, index) => (
+              <button
+                key={label}
+                onClick={() => setActiveDay(index)}
+                className={cn(
+                  "flex flex-1 flex-col items-center gap-2 rounded-lg py-12 transition-colors duration-200",
+                  activeDay === index
+                    ? "bg-orange-500 text-white"
+                    : "bg-gray-100 text-gray-500 hover:bg-gray-200",
+                )}
+              >
+                <span className="text-body3b mobile:text-caption1b">{label}</span>
+                <span className="text-caption2r opacity-90">{date}</span>
+              </button>
+            ))}
+          </div>
+          <YouTubeLazyEmbed
+            key={activeStream.videoId}
+            videoId={activeStream.videoId}
+            title={`2026 광탈페 예선 ${activeStream.label} 실시간 중계`}
+          />
+        </div>
+
+        <div className="flex flex-col gap-16">
+          <h3 className="text-body2b mobile:text-body3b text-black">예선 공연 순서</h3>
+          <Image
+            src="/images/preliminary-lineup.jpg"
+            alt="2026 광탈페 예선 공연 순서"
+            width={3508}
+            height={2480}
+            className="w-full h-auto rounded-lg border border-gray-100"
+            sizes="(max-width: 768px) 100vw, 70vw"
+          />
+        </div>
+      </div>
+    </section>
+  );
+};
+
+export default PreliminaryLiveSection;
